Move ViewScopeEventListener to appropriate package.
